sosdb_auth user_id default random ID fix
Issue: The default value of `SosdbUser.user_id` was set to
`randint(1, 1000)`. I believe the intention was to generate a new
random ID if the `user_id` value was not given. However, what really
happened when `makemigrations` is called is that `randint()` is
evaluated, and our default is a (random) fixed number. The effects of
this are:

1) When `makemigrations` is called again SosdbUser is changed dueto
   the default value change caused by `randint()` evaluation.

2) The default value of the new object is not randomly generated when
   the object is created. If we create 2 new users without giving
   `user_id`, they will have the same default (random but fixed) ID
   generated when `makemigrations` is called.

I believe this is not what we want.

Fix: a `callable` is supplied to `SosdbUser.user_id` default so that
the default value is evaluated at object creation time.